fucking wip refactoring mothafocka
