+++ scripts/configs.js
@@ -1,0 +1,7 @@
+(
+    function() {
+        simulator.configs = {
+            URL_API_SIMULATOR: 'https://easynvestsimulatorcalcapi.azurewebsites.net/calculator/simulate'
+        }
+    }
+)()
