+++ scripts/index.js
@@ -1,0 +1,1 @@
+const simulator = {}
